requirements.txt
openai-whisper >= 20231117
pyyaml >= 6.0
watchdog >= 3.0.0
requests >=2.31.0
#CPU用
torch